+++ src/main/java/seedu/taskman/logic/commands/DoCommand.java
@@ -1,12 +1,17 @@
 package seedu.taskman.logic.commands;
 
 import seedu.taskman.commons.exceptions.IllegalValueException;
+import seedu.taskman.logic.parser.CommandParser;
 import seedu.taskman.model.tag.Tag;
 import seedu.taskman.model.tag.UniqueTagList;
 import seedu.taskman.model.event.*;
 
 import java.util.HashSet;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static seedu.taskman.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 /**
  * Adds a Task to the task man.
@@ -23,6 +28,12 @@ public class DoCommand extends Command {
 
     public static final String MESSAGE_SUCCESS = "New task added: %1$s";
     public static final String MESSAGE_DUPLICATE_EVENT = "This task already exists in TaskMan";
+    private static final Pattern TASK_DO_ARGS_FORMAT = // '/' forward slashes are reserved for delimiter prefixes
+            Pattern.compile("" + CommandParser.Argument.TITLE
+                    + CommandParser.Argument.DEADLINE
+                    + CommandParser.Argument.SCHEDULE
+                    + CommandParser.Argument.FREQUENCY
+                    + CommandParser.Argument.TAG); // variable number of tags
 
     private final Task toAdd;
 
@@ -31,7 +42,7 @@ public class DoCommand extends Command {
      *
      * @throws IllegalValueException if any of the raw values are invalid
      */
-    public DoCommand(String title, String deadline, String schedule, String frequency, Set<String> tags)
+    private DoCommand(String title, String deadline, String schedule, String frequency, Set<String> tags)
             throws IllegalValueException {
         final Set<Tag> tagSet = new HashSet<>();
         for (String tagName : tags) {
@@ -50,6 +61,24 @@ public class DoCommand extends Command {
                     ? null
                     : new Frequency(frequency)
         );
+    }
+
+    public static Command prepareDo(String args) {
+        final Matcher matcher = TASK_DO_ARGS_FORMAT.matcher(args.trim());
+        if (!matcher.matches()) {
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, MESSAGE_USAGE));
+        }
+        try {
+            return new DoCommand(
+                    matcher.group("title"),
+                    matcher.group("deadline"),
+                    matcher.group("schedule"),
+                    matcher.group("frequency"),
+                    getTagsFromArgs(matcher.group("tagArguments"))
+            );
+        } catch (IllegalValueException ive) {
+            return new IncorrectCommand(ive.getMessage());
+        }
     }
 
     @Override
